fix: Fix Kotlin compilation errors in BitcoinPuzzleScreen and BitcoinPuzzleViewModel
- Use items() instead of forEach in LazyColumn for Composable context
- Fix BigInteger constructor to use valueOf() for Long values
- Fix BigInteger division to use BigInteger.valueOf(2)

+++ jackscanner/app/src/main/java/com/jackscanner/ui/screens/puzzle/BitcoinPuzzleViewModel.kt
@@ -159,7 +159,7 @@ class BitcoinPuzzleViewModel @Inject constructor() : ViewModel() {
                     SearchMethod.RANDOM -> {
                         val range = end - start
                         if (range > BigInteger.ZERO && range < BigInteger(Long.MAX_VALUE.toString())) {
-                            start + BigInteger(Random.nextLong(range.toLong()))
+                            start + BigInteger.valueOf(Random.nextLong(range.toLong()))
                         } else start
                     }
                     SearchMethod.FIBONACCI -> {
@@ -170,7 +170,7 @@ class BitcoinPuzzleViewModel @Inject constructor() : ViewModel() {
                         } else start
                     }
                     SearchMethod.BINARY_SEARCH -> {
-                        if (count % 2 == 0L) (start + end) / 2 else start
+                        if (count % 2 == 0L) (start + end) / BigInteger.valueOf(2) else start
                     }
                 }
